Implement TUI using crossterm, and fix a bug when getting display position with no content.
